==> Movies/1_Data_Acquisition_and_Cleaning.ipynb <==
# Data Acquisition and Cleaning
<br>
Predictive modelling is highly dependent on the integrity of the data. For my part for the data acquisition, I first obtained data on the Oscar from 1927 -2020 from a reliable source on Kaggle as a starting point. Next I obtained the most up-to-date and reliable from TMDB using the TMDB API and also other useful data from the OMDB API which can be accessed using the IMDB ID which can be retrieve from the TMDB API
<br>
The aim of our project or more specifically our problem defination is to predict the winners of The 93rd Academy Awards which will be held later this month on 26th April 2021.
<br>
The first phase of the project is the Data Acquistion which will be dividied in a couple of steps:
<ol>1. Obtain Oscar Data from Kaggle and to scrape the data and only focus on the big 5 award categories.</ol>
<ol>2. Using the data collected, retrieve data from TMDB API.</ol>
<ol>3. Using the IMDB ID retrieved from the TMDB API, to collect more data such as total nomiations, total awards won, imdb rating, rotten 

# Importing neccesary libraries 
from tmdbv3api import TMDb, Movie, Discover, Genre
from tmdbv3api import Account
from tmdbv3api import Authentication
import numpy as np
import random
import pandas as pd
from collections import defaultdict
import requests as rq
import re
import traceback
import datetime
from collections import Counter
import omdb

##TMDB details
##Account details for getting movie reccomendations
##Run this block only once during a session
tmdb = TMDb()
tmdb.api_key = '0921b0cce35c0b2ec8b874614d1d0b47'
tmdb.language = 'en'

USERNAME = "cybercat94"
PASSWORD = "Wave1994@"
auth = Authentication(username=USERNAME, password=PASSWORD)
account = Account()
details = account.details()

print("You are logged in as %s. Your account ID is %s." % (details.username, details.id))
print("This session expires at: %s" % auth.expires_at)

## Scrapping and cleaning oscar_award_data.csv
We will start by scrapping the dataset to meet our needs and requirements.

oscar_df = pd.read_csv('data/the_oscar_award.csv')
oscar_df

Renaming columns for better readabality and querying data from the year 1960 to 2020.

oscar_df= oscar_df.rename({'category':'oscar_category', 'name':'nominee', 'winner':'oscar_win'}, axis=1)
oscar_df.drop(['ceremony'], axis=1, inplace=True)
oscar_df = oscar_df.query('year_film >=1960')

Next is to get the total counts of nominations for each film.

no_of_wins_df = oscar_df[oscar_df['oscar_win']==True].groupby('film').size().reset_index(name='no_oscar_wins')
no_of_wins_df


no_of_nomination_df = oscar_df.groupby('film').size().reset_index(name='no_oscar_nominations')
df = pd.merge(no_of_wins_df, no_of_nomination_df, on=['film'], how='right').fillna(0)
pd.options.display.float_format = '{:,.0f}'.format
df

For this project we will be focusing on major awards only, the so called "big 5".
<br>Best Actor
<br>Best Actress
<br>Best Director
<br>Best Picture
<br>Best Adapted Screenplay/Best Original Screenplay

<br>As the title of these awards kept changing through the years, we will have to filter through the data.


oscar_df = oscar_df.replace(to_replace={'oscar_category' : {'ACTOR IN A LEADING ROLE': 'Best Actor',
                                                            'ACTOR': 'Best Actor',
                                                            'ACTRESS': 'Best Actress',
                                                            'ACTRESS IN A LEADING ROLE': 'Best Actress',
                                                            'DIRECTING' : 'Best Director',
                                                            'BEST PICTURE' : 'Best Picture',
                                                            'BEST MOTION PICTURE' : 'Best Picture',
                                                            'WRITING (Adapted Screenplay)' : 'Best Adapted Screenplay',
                                                            'WRITING (Screenplay Adapted from Other Material)' : 'Best Adapted Screenplay',
                                                            'WRITING (Screenplay Based on Material from Another Medium)' : 'Best Adapted Screenplay',
                                                            'WRITING (Screenplay--based on material from another medium)' : 'Best Adapted Screenplay',
                                                            'WRITING (Screenplay Based on Material Previously Produced or Published)' : 'Best Adapted Screenplay',
                                                            'WRITING (Original Screenplay)' : 'Best Original Screenplay',
                                                            'WRITING (Screenplay Written Directly for the Screen)' : 'Best Original Screenplay',
                                                            'WRITING (Story and Screenplay--based on factual material or material not previously published or produced)' : 'Best Original Screenplay',
                                                            'WRITING (Story and Screenplay--written directly for the screen)' : 'Best Original Screenplay',
                                                            'WRITING (Screenplay Written Directly for the Screen--based on factual material or on story material not previously published or produced)' : 'Best Original Screenplay',
                                                            'WRITING (Story and Screenplay--based on material not previously published or produced)' : 'Best Original Screenplay'},}, value=None)

oscar_df = oscar_df.query('oscar_category in ["Best Actor", "Best Actress", "Best Director", "Best Picture", "Best Adapted Screenplay","Best Original Screenplay"]')
oscar_df = pd.merge(oscar_df, df, on=['film'], how='left')
oscar_df = oscar_df.reset_index()
oscar_df.drop(['index'], axis=1, inplace=True)
oscar_df

# TMDB API

## Get Movie Information from TMDB api
This function uses TMDB api(DISCOVER), to retrieve all the movies with the constraint year.
<br> It will retrieve basic details such as id, title, list of genre ids, vote average and lastly vote count.
<br> We will later retrieve more information using the getDetails function to get details like casts, crews and etc.

def get_movies_by_year(year):

    movie = Movie()

    # initialiZe empty dictionary
    _movie_dict = {"tmdb_id":[], 
                   "year":[], 
                   "title":[], 
                   "genre":[], 
                   "tmdb_vote_average":[], 
                   "tmdb_vote_count":[]};
    
    print(f"In {year}")
    
    for i in range(1,501):
        discover = Discover()
        movie = discover.discover_movies({
            'year': year,
            'page': i,
            'certification_country': 'US'
        })
        
        for m in movie:
            _movie_dict["tmdb_id"].append(m.id)
            _movie_dict["year"].append(year)
            _movie_dict["title"].append(m.title)
            _movie_dict["genre"].append(m.title)
            _movie_dict["tmdb_vote_average"].append(m.title)
            _movie_dict["tmdb_vote_count"].append(m.title)

        if i == int(discover.total_pages):
            break

    return _movie_dict

Function to translate the genre id

def translate_genre_id(g):
    for i in range(0,len(g)):
        if g[i] == 28:
            g[i] = 'Action'
        elif g[i] == 12:
            g[i] = 'Adventure'
        elif g[i] == 16:
            g[i] = 'Animation'
        elif g[i] == 35:
            g[i] = 'Comedy'
        elif g[i] == 80:
            g[i] = 'Crime'
        elif g[i] == 99:
            g[i] = 'Documentary'
        elif g[i] == 18:
            g[i] = 'Drama'
        elif g[i] == 10751:
            g[i] = 'Family'
        elif g[i] == 14:
            g[i] = 'Fantasy'
        elif g[i] == 36:
            g[i] = 'History'
        elif g[i] == 27:
            g[i] = 'Horror'
        elif g[i] == 10402:
            g[i] = 'Music'
        elif g[i] == 9648:
            g[i] = 'Mystery'
        elif g[i] == 10749:
            g[i] = 'Romance'
        elif g[i] == 878:
            g[i] = 'Science Fiction'
        elif g[i] == 10770:
            g[i] = 'TV Movie'
        elif g[i] == 53:
            g[i] = 'Thriller'
        elif g[i] == 10752:
            g[i] = 'War'
        elif g[i] == 37:
            g[i] = 'Western'
    return g

Function to retrieve movie details with movie name and year

def search_movie_by_name_year(details_dict):
    
    oscar_dict = {"tmdb_id":[], 
                  "year_film":[], 
                  "year_ceremony":[], 
                  "film":[], 
                  "oscar_category":[], 
                  "nominee":[],
                  "no_oscar_wins":[], 
                  "no_oscar_nominations":[], 
                  "oscar_win":[],
                  "genre_id":[], 
                  "tmdb_vote_average":[],
                  "tmdb_vote_count":[]};
    
    movie = Movie()
    for item in details_dict:
        s = movie.search(item["film"])
        size = len(s)
        for i in range(0, size):
            result = s[i]
            try:
                if (result.release_date[0:4] == str(item["year_film"]) or result.release_date[0:4] == str(item["year_film"]-1) or result.release_date[0:4] == str(item["year_film"]-2)):
                    oscar_dict["tmdb_id"].append(result.id)
                    oscar_dict["year_film"].append(item["year_film"])
                    oscar_dict["year_ceremony"].append(item["year_ceremony"])
                    oscar_dict["oscar_category"].append(item["oscar_category"])
                    oscar_dict["film"].append(item["film"])
                    oscar_dict["nominee"].append(item["nominee"])
                    oscar_dict["no_oscar_wins"].append(item["no_oscar_wins"])
                    oscar_dict["no_oscar_nominations"].append(item["no_oscar_nominations"])
                    oscar_dict["oscar_win"].append(item["oscar_win"])
                    oscar_dict["genre_id"].append(result.genre_ids)
                    oscar_dict["tmdb_vote_average"].append(int((result.vote_average)*10))
                    oscar_dict["tmdb_vote_count"].append(result.vote_count)
                    break
            except NameError:
                print("no release date")
            except:
                print("Error can't find movie")

    return oscar_dict

details_dict = oscar_df.to_dict('r')
df = pd.DataFrame(search_movie_by_name_year(details_dict))

df

Function to retrieve top 4 casts 

def get_top_4_casts(creds):
    
    cast = (creds["cast"])
    cast_list = []
    
    for i in range(0,4):
        try:
            cast_list.append(cast[i].name)
        except:
            print("OUT OF RANGE")
        
    return cast_list

Function to retrieve total popularity score of top 4 casts 

def get_top_4_casts_popularity_score(creds):
    
    cast = (creds["cast"])
    popularity_score = 0

    for i in range(0,4):
        try:
            popularity_score += cast[i].popularity
        except:
            print("OUT OF RANGE")
        
    return popularity_score

Function to get total casts score

def get_casts_popularity_score(creds):
    
    cast = (creds["cast"])
    popularity_score = 0

    for c in cast:
        try:
            popularity_score += c.popularity
        except:
            print("OUT OF RANGE")
        
    return popularity_score

Function to retrieve crew based on positions

def get_crew_position(creds, job_id):
    crew = (creds["crew"])
    producer_list = []
    screenplay_list = []
    for c in crew:
        try:
            if c.job == 'Director' and job_id == 0:
                director = c.name
            elif c.job == 'Producer'and job_id == 1:
                producer_list.append(c.name)
            elif c.job == 'Screenplay'and job_id == 2:
                screenplay_list.append(c.name)
        except:
             print("OUT OF RANGE")
    if job_id == 0:
        return director
    elif job_id == 1:
        return producer_list
    elif job_id == 2:
        return screenplay_list

Function to get crew score

def get_crew_score(creds):
    crew = (creds["crew"])
    popularity_score = 0
    for c in crew:
        try:
            popularity_score += c.popularity
        except:
            print("OUT OF RANGE")
        
    return popularity_score

Function to get more movie details

def get_movie_details(oscar_dict):
    
    movie = Movie()
    
    movie_detials_dict = {"tmdb_id":[],
                           "imdb_id":[],
                           "year_film":[],
                           "year_ceremony":[], 
                           "film":[], 
                           "oscar_category":[], 
                           "nominee":[],
                           "oscar_win":[],
                           "no_oscar_wins":[], 
                           "no_oscar_nominations":[], 
                           "genre_id":[],
                           "staring_casts":[],
                           "casts_popularity":[],
                           "director":[],
                           "producers":[],
                           "screenplay":[],
                           "crew_popularity":[],
                           "budget":[],
                           "revenue":[],
                           "runtime":[],
                           "tmdb_vote_average":[], 
                           "tmdb_vote_count":[]};
                        
    for item in oscar_dict:
        if item["tmdb_id"] != 0:
            m = movie.details(item["tmdb_id"])
            movie_detials_dict["tmdb_id"].append(item["tmdb_id"])
            movie_detials_dict["imdb_id"].append(m.imdb_id)
            movie_detials_dict["year_film"].append(item["year_film"])
            movie_detials_dict["year_ceremony"].append(item["year_ceremony"])
            movie_detials_dict["film"].append(item["film"])
            movie_detials_dict["oscar_category"].append(item["oscar_category"])
            movie_detials_dict["nominee"].append(item["nominee"])
            movie_detials_dict["oscar_win"].append(item["oscar_win"])
            movie_detials_dict["no_oscar_wins"].append(item["no_oscar_wins"])
            movie_detials_dict["no_oscar_nominations"].append(item["no_oscar_nominations"])
            movie_detials_dict["genre_id"].append(item["genre_id"])
            movie_detials_dict["staring_casts"].append(get_top_4_casts(m.casts))
            movie_detials_dict["casts_popularity"].append(get_casts_popularity_score(m.casts))
            movie_detials_dict["director"].append(get_crew_position(m.casts, 0))
            movie_detials_dict["producers"].append(get_crew_position(m.casts, 1))
            movie_detials_dict["screenplay"].append(get_crew_position(m.casts, 2))
            movie_detials_dict["crew_popularity"].append(get_crew_score(m.casts))
            movie_detials_dict["budget"].append(m.budget)
            movie_detials_dict["revenue"].append(m.revenue)
            movie_detials_dict["runtime"].append(m.runtime)
            movie_detials_dict["tmdb_vote_average"].append(item["tmdb_vote_average"])
            movie_detials_dict["tmdb_vote_count"].append(item["tmdb_vote_count"])
                          
    return movie_detials_dict


oscar_dict = df.to_dict('r')
final_df = pd.DataFrame(get_movie_details(oscar_dict))

final_df

# OMDB API
Next, we would like to get more information on the movie using the IMDB ID that we obtained using the TMDB API.
<br>We want to get information like:
<ol>1. Metascore </ol>
<ol>2. Imdb score </ol>
<ol>3. Rotten score </ol>
<ol>4. other nominations and wins </ol>


omdb.set_default('apikey', '4a7f4869')

def get_omdb_details(oscar_dict):
    
    omdb_detials_dict = {"imdb_id":[],
                         "oscar_category":[],
                         "nominee":[],
                         "imdb_rating":[],
                         "imdb_votes":[],
                         "rotten_tomatoes_rating":[],
                         "metascore":[],
                         "total_wins":[],
                         "total_nominations":[]};
    for item in oscar_dict:
        try:
            c = omdb.imdbid(item['imdb_id'])
            words = (c['awards'].split())
            imdb_rating = int((c['imdb_rating']).replace(".",""))
            imdb_votes = int((c['imdb_votes']).replace(',',''))
            if len(c['ratings']) > 1:
                rotten_tomatoes_rating = int((c['ratings'][1]['value'])[0:2])
            else: 
                rotten_tomatoes_rating = "N/A"
            if c['metascore'] != "N/A":
                metascore = int(c['metascore'])
            else:
                metascore = c['metascore']
            for key in words:
                try:
                    if key == "wins":
                        total_wins = int(words[(words.index(key)-1)])
                    if key == "nominations.":
                        total_nominations = int(words[(words.index(key)-1)])
                except:
                    print("not found")
            omdb_detials_dict["imdb_id"].append(item['imdb_id'])
            omdb_detials_dict["oscar_category"].append(item['oscar_category'])
            omdb_detials_dict["nominee"].append(item['nominee'])
            omdb_detials_dict["imdb_rating"].append((imdb_rating))
            omdb_detials_dict["imdb_votes"].append((imdb_votes))
            omdb_detials_dict["rotten_tomatoes_rating"].append((rotten_tomatoes_rating))
            omdb_detials_dict["metascore"].append((metascore))
            omdb_detials_dict["total_wins"].append(int(total_wins))
            omdb_detials_dict["total_nominations"].append(int(total_nominations))
        except:
            print("error")
            print(c['title'])
    return omdb_detials_dict

imdb = oscar_details.to_dict('r')
omdb_df = pd.DataFrame(get_omdb_details(imdb))

omdb_df

final_movie_df = pd.merge(final_df, omdb_df, how='left', on=["imdb_id", "oscar_category", "nominee"])
final_movie_df

db_path='Final_Movie_Data''.csv'
final_movie_df.to_csv(db_path)
print("Exported to local folder:",db_path)

# References
https://www.kaggle.com/unanimad/the-oscar-award
<br>http://www.omdbapi.com/
<br>https://developers.themoviedb.org/3
<br>https://pypi.org/project/tmdbv3api/
