==> Movies/1_Data_cleaning_continued.ipynb <==
## Further data cleaning
Objectives:
- Convert string representation of lists into list
    - Assign each element of new list to a standalone column
    - Number of columns to assign depends on the average length of lists in existing column  
  
- 
- Reorder the DataFrame to aid in exploratory analysis and machine learning

## Stuff to do:
- Convert all fields to the best possible data type
- Remove unwanted headers: budget, revenue, unnamed, runtime
- Convert string representation of lists to list. Assign 3 new columns for each list
    - Headers in question are producers, screenplay, starring_casts, genre_id (4 headers)

import numpy as np
import pandas as pd
import seaborn as sb
import matplotlib.pyplot as plt
from collections import Counter
import statistics as stat

## Own functions
import _functions_1115 as func

## Import dirty data
data_raw = pd.read_csv('./data/Final_Movie_Data_Cleaned.csv')
data_raw.info(10)
##data_raw['film']

## from now on, data_cleaning will be used as a temp variable when changing the contents of data_raw
## data_cleaning will be continuously overwritten until arriving at final cleaned data
data_cleaning = data_raw.copy(deep=True)

## remove unwanted headers - those that serve no purpose in movie identification/machine learning
## budget and revenue might be filled with mean, so might be still usable
data_cleaning.drop(['Unnamed: 0','Unnamed: 0.1','runtime','nominee'], axis=1, inplace=True)

## convert to the best fit datatype
data_cleaning = data_cleaning.convert_dtypes()
data_cleaning.info()

## rename some headers 
data_cleaning.rename({
    'year_film':'year',
    'oscar_category':'oscar_cat',
    'no_oscar_wins':'oscar_win_count',
    'no_oscar_nominations':'oscar_nominations',
    'staring_casts':'cast',
    'casts_popularity':'cast_popularity',
    'rotten_tomatoes_rating':'rotten_tomatoes'
}, axis=1, inplace=True)
data_cleaning.info()

## Checking the number of cols needed for each of the following fields
for col in ['genre_id','cast','producers','screenplay']:
    #check the number of genre_ids in each row
    temp_list = []
    
    ## append length of each list 
    for entry in data_cleaning[col]:
        #print(func.strtolist(entry))
        temp_list.append(len(func.strtolist(entry)))

    print('''Stats for number of genres in column '%s':
    Mean:   %d
    Median: %d
    Mode:   %d
    Min:    %d
    Max:    %d\n''' % 
    (col, stat.mean(temp_list), stat.median(temp_list), stat.mode(temp_list), min(temp_list), max(temp_list)))

## creating new col for each item in list, based on the cumuative stats per col above
## taking the median no. of items in a list
new_headers = {
    'genre_id':['genre_id_0','genre_id_1'], 
    'cast':['cast_0','cast_1','cast_2','cast_3'],
    'producers':['producer_0','producer_1'],
    'screenplay':['screenplay_0']
}

## replacing cols with string representation of list to an actual list
for item in new_headers:
    data_cleaning[item] = [func.strtolist(data_cleaning[item][index]) for index in range(len(data_cleaning))]

## WORKING
## What this does is take each existing col in the new_headers dictionary, and expands out to multiple sub columns
## This is because the existing col contains a list that is not useful when doing machine learning

## look for current col that exists in 
for currheader in new_headers:
    ## look for target subcols, check if they exist first
    if currheader in data_cleaning:
        ## perform addition of new cols from currheader list containing only 1 entry, NAN if list index out of range
        for subcol in range(len(new_headers[currheader])):
            
            tmp_list = []
            for i in range(len(data_cleaning)):
                ## checks if list index still in range
                try:
                    tmp_list.append(data_cleaning[currheader][i][subcol])
                ## adds NAN if out of range
                except:
                    tmp_list.append(None)

            data_cleaning[new_headers[currheader][subcol]] = tmp_list
            
data_cleaning = data_cleaning.convert_dtypes()

## removing even more cols, including the cols containing lists of str/int (cause these elements are added to multiple subcols already)
data_cleaning.drop([
    'year_ceremony', 'genre_id','cast','producers','screenplay'
], axis=1, inplace=True)

data_cleaning.info()

## reordering the cols for clarity
#col_list = data_cleaning.columns.tolist()
#display(len(data_cleaning.columns))
#display(col_list)

col_list_reorder = [
    'tmdb_id',
    'imdb_id',
    'film',
    'year',
    'budget', ## not inside?
    'revenue', ## not inside?
    'oscar_cat',
    'oscar_win',
    'oscar_nominations',
    'oscar_win_count',
    'total_wins',
    'total_nominations',
    'cast_popularity',
    'crew_popularity',
    'tmdb_vote_average',
    'tmdb_vote_count',
    'imdb_rating',
    'imdb_votes',
    'rotten_tomatoes',
    'metascore',
    'director',
    'genre_id_0',
    'genre_id_1',
    'cast_0',
    'cast_1',
    'cast_2',
    'cast_3',
    'producer_0',
    'producer_1',
    'screenplay_0'
]
data_cleaning = data_cleaning[col_list_reorder]

data_cleaning.info()

## Skip this cell
## converting data types to appropriate ones
data_cleaning = data_cleaning.astype({
    'oscar_cat':'object',
    'director':'object',
    'genre_id_0':'object',
    'genre_id_1':'object',
    'starring_0':'object',
    'starring_1':'object',
    'starring_2':'object',
    'starring_3':'object',
    'producer_0':'object',
    'producer_1':'object',
    'screenplay_0':'object'
})

data_cleaning.info()

## finally assign variable data_clean
data_clean = data_cleaning.copy(deep=True)
## only perform this command if there are changes to be made to the csv file
#data_clean.to_csv('./data/Final_Movie_Data_clean.csv')

---
### Test cells below
---

test = data_cleaning.copy(deep=True)
test = test[['genre_id','producers']]
#test = test.head(100)
## replacing by this method works, using list comprehension
test['genre_id'] = [func.strtolist(test['genre_id'][index]) for index in range(len(test))]
test['producers'] = [func.strtolist(test['producers'][index]) for index in range(len(test))]
display(test)

test_list = []
for i in range(len(test)):
    try:
        test_list.append(test['genre_id'][i][0])
    except:
        test_list.append(None)

test['genre_id_0'] = test_list


## do this last
test.convert_dtypes()
#test['genre_id'][3][0]
if 'genre_id' in test:
    print(1)

## WORKING
## look for current col that exists in 
for currheader in new_headers:
    ## look for target subcols, check if they exist first
    if currheader in test:
        for subcol in range(len(new_headers[currheader])):
            
            tmp_list = []
            for i in range(len(test)):
                try:
                    tmp_list.append(test[currheader][i][subcol])
                except:
                    tmp_list.append(None)

            test[new_headers[currheader][subcol]] = tmp_list
            
test = test.convert_dtypes()
test

## comparing data before and after cleaning pass
display(data_raw.info())
display(data_clean.info())

newdata = pd.read_csv('./data/Final_Movie_Data_Cleaned.csv')
display(newdata.info())
display(data_clean.info())
