==> Movies/2_Data_Cleaning.ipynb <==
# Data Cleaning
With missing values in the budget and revenue column. Although this might not be important in predicting this years winners due to covid-19 which heavily disrupted the film industry with movite thertres. I believe this can present some valuable insights. Hence, I'll be repolacing all these NULL values wiht the mean of budget and revenue according to the award categories

import numpy as np
import pandas as pd

movie_df =  pd.read_csv("Final_Movie_Data.csv")
movie_df.head()

movie_df.oscar_category.unique()

print("Total Records with OSCAR Category of Best Actor" , movie_df.oscar_category[movie_df.oscar_category == 'Best Actor'].count())
print("Total Records with OSCAR Category of Best Actress",movie_df.oscar_category[movie_df.oscar_category == 'Best Actress'].count())
print("Total Records with OSCAR Category of Best Director", movie_df.oscar_category[movie_df.oscar_category == 'Best Director'].count())
print("Total Records with OSCAR Category of Best Picture",movie_df.oscar_category[movie_df.oscar_category == 'Best Picture'].count())
print("Total Records with OSCAR Category of Best Adapted Screenplay",movie_df.oscar_category[movie_df.oscar_category == 'Best Adapted Screenplay'].count())
print("Total Records with OSCAR Category of Best Original Screenplay", movie_df.oscar_category[movie_df.oscar_category == 'Best Original Screenplay'].count())


### Budget Mean For Different Categories

budget_best_actor_mean = movie_df.budget[movie_df.oscar_category == 'Best Actor'].mean()
budget_best_actoress_mean = movie_df.budget[movie_df.oscar_category == 'Best Actress'].mean()
budget_best_director_mean = movie_df.budget[movie_df.oscar_category == 'Best Director'].mean()
budget_best_picture_mean = movie_df.budget[movie_df.oscar_category == 'Best Picture'].mean()
budget_best_adapted_screenplay_mean = movie_df.budget[movie_df.oscar_category == 'Best Adapted Screenplay'].mean()
budget_best_original_screenplay_mean = movie_df.budget[movie_df.oscar_category == 'Best Original Screenplay'].mean()

print("Budget Mean of Best Actor ", budget_best_actor_mean)
print("Budget Mean of Best Actoress ",budget_best_actoress_mean)
print("Budget Mean of Best Director ",budget_best_director_mean)
print("Budget Mean of Best Picture ",budget_best_picture_mean)
print("Budget Mean of Best Adapted Screenplay ",budget_best_adapted_screenplay_mean)
print("Budget Mean of Best Original Screenplay ",budget_best_original_screenplay_mean)

### Revenue Mean For Different Categories

revenue_best_actor_mean = movie_df.revenue[movie_df.oscar_category == 'Best Actor'].mean()
revenue_best_actoress_mean = movie_df.revenue[movie_df.oscar_category == 'Best Actress'].mean()
revenue_best_director_mean = movie_df.revenue[movie_df.oscar_category == 'Best Director'].mean()
revenue_best_picture_mean = movie_df.revenue[movie_df.oscar_category == 'Best Picture'].mean()
revenue_best_adapted_screenplay_mean = movie_df.revenue[movie_df.oscar_category == 'Best Adapted Screenplay'].mean()
revenue_best_original_screenplay_mean = movie_df.revenue[movie_df.oscar_category == 'Best Original Screenplay'].mean()

print("Revenue Mean of Best Actor ", revenue_best_actor_mean)
print("Revenue Mean of Best Actoress ",revenue_best_actoress_mean)
print("Revenue Mean of Best Director ",revenue_best_director_mean)
print("Revenue Mean of Best Picture ",revenue_best_picture_mean)
print("Revenue Mean of Best Adapted Screenplay ",revenue_best_adapted_screenplay_mean)
print("Revenue Mean of Best Original Screenplay ",revenue_best_original_screenplay_mean)

### Total Null Values in Budget For Different Categories

## Null Values in Budget For Different Categories

print("Total Null Values in Budget For Best Actor " ,movie_df.budget[movie_df.oscar_category == 'Best Actor'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Actress ", movie_df.budget[movie_df.oscar_category == 'Best Actress'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Director ",movie_df.budget[movie_df.oscar_category == 'Best Director'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Picture ",movie_df.budget[movie_df.oscar_category == 'Best Picture'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Adapted Screenplay ",movie_df.budget[movie_df.oscar_category == 'Best Adapted Screenplay'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Original Screenplay ",movie_df.budget[movie_df.oscar_category == 'Best Original Screenplay'][movie_df.budget == 0].count())

filter1 = ( (movie_df.oscar_category == 'Best Actor') &  (movie_df.budget == 0))
filter2 = ( (movie_df.oscar_category == 'Best Actress') &  (movie_df.budget == 0))
filter3 = ( (movie_df.oscar_category == 'Best Director') &  (movie_df.budget == 0))
filter4 = ( (movie_df.oscar_category == 'Best Picture') &  (movie_df.budget == 0))
filter5 = ( (movie_df.oscar_category == 'Best Adapted Screenplay') &  (movie_df.budget == 0))
filter6 = ( (movie_df.oscar_category == 'Best Original Screenplay') &  (movie_df.budget == 0))

movie_df.loc[filter1,'budget'] =  budget_best_actor_mean
movie_df.loc[filter2,'budget'] =  budget_best_actoress_mean
movie_df.loc[filter3,'budget'] =  budget_best_director_mean
movie_df.loc[filter4,'budget'] =  budget_best_picture_mean
movie_df.loc[filter5,'budget'] =  budget_best_adapted_screenplay_mean
movie_df.loc[filter6,'budget'] =  budget_best_original_screenplay_mean

### Null Values in Budget For Different Categories After Imputation

## Null Values in Budget For Different Categories After Imputation

print("Total Null Values in Budget For Best Actor " ,movie_df.budget[movie_df.oscar_category == 'Best Actor'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Actress ", movie_df.budget[movie_df.oscar_category == 'Best Actress'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Director ",movie_df.budget[movie_df.oscar_category == 'Best Director'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Picture ",movie_df.budget[movie_df.oscar_category == 'Best Picture'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Adapted Screenplay ",movie_df.budget[movie_df.oscar_category == 'Best Adapted Screenplay'][movie_df.budget == 0].count())
print("Total Null Values in Budget For Best Original Screenplay ",movie_df.budget[movie_df.oscar_category == 'Best Original Screenplay'][movie_df.budget == 0].count())

### Total Null Values in Revenue For Different Categories

## Null Values in Revenue For Different Categories

print("Total Null Values in Revenue For Best Actor " ,movie_df.revenue[movie_df.oscar_category == 'Best Actor'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Actress ", movie_df.revenue[movie_df.oscar_category == 'Best Actress'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Director ",movie_df.revenue[movie_df.oscar_category == 'Best Director'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Picture ",movie_df.revenue[movie_df.oscar_category == 'Best Picture'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Adapted Screenplay ",movie_df.revenue[movie_df.oscar_category == 'Best Adapted Screenplay'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Original Screenplay ",movie_df.revenue[movie_df.oscar_category == 'Best Original Screenplay'][movie_df.revenue == 0].count())

filter1 = ( (movie_df.oscar_category == 'Best Actor') &  (movie_df.revenue == 0))
filter2 = ( (movie_df.oscar_category == 'Best Actress') &  (movie_df.revenue == 0))
filter3 = ( (movie_df.oscar_category == 'Best Director') &  (movie_df.revenue == 0))
filter4 = ( (movie_df.oscar_category == 'Best Picture') &  (movie_df.revenue == 0))
filter5 = ( (movie_df.oscar_category == 'Best Adapted Screenplay') &  (movie_df.revenue == 0))
filter6 = ( (movie_df.oscar_category == 'Best Original Screenplay') &  (movie_df.revenue == 0))

movie_df.loc[filter1,'revenue'] =  revenue_best_actor_mean
movie_df.loc[filter2,'revenue'] =  revenue_best_actoress_mean
movie_df.loc[filter3,'revenue'] =  revenue_best_director_mean
movie_df.loc[filter4,'revenue'] =  revenue_best_picture_mean
movie_df.loc[filter5,'revenue'] =  revenue_best_adapted_screenplay_mean
movie_df.loc[filter6,'revenue'] =  revenue_best_original_screenplay_mean

### Null Values in Revenue For Different Categories After Imputation

print("Total Null Values in Revenue For Best Actor " ,movie_df.revenue[movie_df.oscar_category == 'Best Actor'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Actress ", movie_df.revenue[movie_df.oscar_category == 'Best Actress'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Director ",movie_df.revenue[movie_df.oscar_category == 'Best Director'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Picture ",movie_df.revenue[movie_df.oscar_category == 'Best Picture'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Adapted Screenplay ",movie_df.revenue[movie_df.oscar_category == 'Best Adapted Screenplay'][movie_df.revenue == 0].count())
print("Total Null Values in Revenue For Best Original Screenplay ",movie_df.revenue[movie_df.oscar_category == 'Best Original Screenplay'][movie_df.revenue == 0].count())

movie_df

movie_df.isnull().any()

There seems to be some null values in rotten_tomatoes_rating and metascore. Let's try and fix that too!

### Rotten Tomatoes Rating Mean For Different Categories

rotten_tomatoes_rating_best_actor_mean = movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Actor'].mean()
rotten_tomatoes_rating_best_actoress_mean = movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Actress'].mean()
rotten_tomatoes_rating_best_director_mean = movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Director'].mean()
rotten_tomatoes_rating_best_picture_mean = movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Picture'].mean()
rotten_tomatoes_rating_best_adapted_screenplay_mean = movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Adapted Screenplay'].mean()
rotten_tomatoes_rating_best_adapted_screenplay_mean = movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Original Screenplay'].mean()

print("Rotten Tomatoes Rating Mean of Best Actor ", rotten_tomatoes_rating_best_actor_mean)
print("Rotten Tomatoes Rating Mean of Actoress ",rotten_tomatoes_rating_best_actoress_mean)
print("Rotten Tomatoes Rating Mean of Best Director ",rotten_tomatoes_rating_best_director_mean)
print("Rotten Tomatoes Rating Mean of Picture ",rotten_tomatoes_rating_best_picture_mean)
print("Rotten Tomatoes Rating Mean of Adapted Screenplay ",rotten_tomatoes_rating_best_adapted_screenplay_mean)
print("Rotten Tomatoes Rating Mean of Original Screenplay ",rotten_tomatoes_rating_best_adapted_screenplay_mean)

### Metascore Mean For Different Categories

metascore_best_actor_mean = movie_df.metascore[movie_df.oscar_category == 'Best Actor'].mean()
metascore_best_actoress_mean = movie_df.metascore[movie_df.oscar_category == 'Best Actress'].mean()
metascore_best_director_mean = movie_df.metascore[movie_df.oscar_category == 'Best Director'].mean()
metascore_best_picture_mean = movie_df.metascore[movie_df.oscar_category == 'Best Picture'].mean()
metascore_best_adapted_screenplay_mean = movie_df.metascore[movie_df.oscar_category == 'Best Adapted Screenplay'].mean()
metascore_best_adapted_screenplay_mean = movie_df.metascore[movie_df.oscar_category == 'Best Original Screenplay'].mean()

print("Metascore Mean of Best Actor ", metascore_best_actor_mean)
print("Metascore Mean of Actoress ",metascore_best_actoress_mean)
print("Metascore Mean of Best Director ",metascore_best_director_mean)
print("Metascore Mean of Picture ",metascore_best_picture_mean)
print("Metascore Mean of Adapted Screenplay ",metascore_best_adapted_screenplay_mean)
print("Metascore Mean of Original Screenplay ",metascore_best_adapted_screenplay_mean)

Interesting! we can see that Rotten Tomatoes have a slightly higher mean score!

### Total Null Values in Rotten Tomatoes Rating For Different Categories

## Null Values in Rotten Tomatoes Rating For Different Categories

print("Total Null Values in Rotten Tomatoes Rating For Best Actor " ,movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Actor'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Actress ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Actress'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Director ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Director'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Picture ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Picture'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Adapted Screenplay ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Adapted Screenplay'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Original Screenplay ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Original Screenplay'].isna().sum())

filter1 = ( (movie_df.oscar_category == 'Best Actor') &  (movie_df.rotten_tomatoes_rating.isna()))
filter2 = ( (movie_df.oscar_category == 'Best Actress') &  (movie_df.rotten_tomatoes_rating.isna()))
filter3 = ( (movie_df.oscar_category == 'Best Director') &  (movie_df.rotten_tomatoes_rating.isna()))
filter4 = ( (movie_df.oscar_category == 'Best Picture') &  (movie_df.rotten_tomatoes_rating.isna()))
filter5 = ( (movie_df.oscar_category == 'Best Adapted Screenplay') &  (movie_df.rotten_tomatoes_rating.isna()))
filter6 = ( (movie_df.oscar_category == 'Best Original Screenplay') &  (movie_df.rotten_tomatoes_rating.isna()))

movie_df.loc[filter1,'rotten_tomatoes_rating'] =  rotten_tomatoes_rating_best_actor_mean
movie_df.loc[filter2,'rotten_tomatoes_rating'] =  rotten_tomatoes_rating_best_actoress_mean
movie_df.loc[filter3,'rotten_tomatoes_rating'] =  rotten_tomatoes_rating_best_director_mean
movie_df.loc[filter4,'rotten_tomatoes_rating'] =  rotten_tomatoes_rating_best_picture_mean
movie_df.loc[filter5,'rotten_tomatoes_rating'] =  rotten_tomatoes_rating_best_adapted_screenplay_mean
movie_df.loc[filter6,'rotten_tomatoes_rating'] =  rotten_tomatoes_rating_best_adapted_screenplay_mean

### Null Values in Rotten Tomatoes Rating  For Different Categories After Imputation

print("Total Null Values in Rotten Tomatoes Rating For Best Actor " ,movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Actor'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Actress ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Actress'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Director ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Director'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Picture ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Picture'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Adapted Screenplay ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Adapted Screenplay'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Original Screenplay ",movie_df.rotten_tomatoes_rating[movie_df.oscar_category == 'Best Original Screenplay'].isna().sum())

### Total Null Values in Metascore For Different Categories

## Null Values in Metascore For Different Categories

print("Total Null Values in Rotten Tomatoes Rating For Best Actor " ,movie_df.metascore[movie_df.oscar_category == 'Best Actor'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Actress ",movie_df.metascore[movie_df.oscar_category == 'Best Actress'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Director ",movie_df.metascore[movie_df.oscar_category == 'Best Director'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Picture ",movie_df.metascore[movie_df.oscar_category == 'Best Picture'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Adapted Screenplay ",movie_df.metascore[movie_df.oscar_category == 'Best Adapted Screenplay'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Original Screenplay ",movie_df.metascore[movie_df.oscar_category == 'Best Original Screenplay'].isna().sum())

filter1 = ( (movie_df.oscar_category == 'Best Actor') &  (movie_df.metascore.isna()))
filter2 = ( (movie_df.oscar_category == 'Best Actress') &  (movie_df.metascore.isna()))
filter3 = ( (movie_df.oscar_category == 'Best Director') &  (movie_df.metascore.isna()))
filter4 = ( (movie_df.oscar_category == 'Best Picture') &  (movie_df.metascore.isna()))
filter5 = ( (movie_df.oscar_category == 'Best Adapted Screenplay') &  (movie_df.metascore.isna()))
filter6 = ( (movie_df.oscar_category == 'Best Original Screenplay') &  (movie_df.metascore.isna()))

movie_df.loc[filter1,'metascore'] =  metascore_best_actor_mean
movie_df.loc[filter2,'metascore'] =  metascore_best_actoress_mean
movie_df.loc[filter3,'metascore'] =  metascore_best_director_mean
movie_df.loc[filter4,'metascore'] =  metascore_best_picture_mean
movie_df.loc[filter5,'metascore'] =  metascore_best_adapted_screenplay_mean
movie_df.loc[filter6,'metascore'] =  metascore_best_adapted_screenplay_mean

### Null Values in Metascore For Different Categories After Imputation

## Null Values in Metascore For Different Categories

print("Total Null Values in Rotten Tomatoes Rating For Best Actor " ,movie_df.metascore[movie_df.oscar_category == 'Best Actor'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Actress ",movie_df.metascore[movie_df.oscar_category == 'Best Actress'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Director ",movie_df.metascore[movie_df.oscar_category == 'Best Director'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Picture ",movie_df.metascore[movie_df.oscar_category == 'Best Picture'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Adapted Screenplay ",movie_df.metascore[movie_df.oscar_category == 'Best Adapted Screenplay'].isna().sum())
print("Total Null Values in Rotten Tomatoes Rating For Best Original Screenplay ",movie_df.metascore[movie_df.oscar_category == 'Best Original Screenplay'].isna().sum())

Let's check if there any other null values that have not yet been handled

movie_df.isnull().any()

Great! It has been handled. Now let's move on to the data exploratory analysis!

db_path='Final_Movie_Data_Cleaned.csv'
movie_df.to_csv(db_path)
print("Exported to local folder:",db_path)
