==> Movies/3_Machine_learning_1.ipynb <==
## List of variables:  
**Main dataset:**
- `data_raw` : data imported locally, needs some touching up
- `data` : data stripped of unneeded information
- `data_ml` : actual data used for machine learning

**For machine learning:**
- `results_<train/test>` : oscar results, in boolean form
- `movinfo_<train/test>` : movie data, as a large dataframe

## Basic Libraries
import numpy as np
import pandas as pd
import seaborn as sb
import matplotlib.pyplot as plt 
sb.set(palette='icefire')

## Machine learning libraries
from sklearn.model_selection import train_test_split
from sklearn.ensemble import AdaBoostClassifier, GradientBoostingClassifier, BaggingClassifier, RandomForestClassifier, ExtraTreesClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn import preprocessing, metrics
import xgboost

from collections import Counter
import statistics as stat
import hashlib ## for converting strings into unique ints
import _functions_1115 as func

## import cleaned data from local dir
data_raw = pd.read_csv('./data/Final_Movie_Data_clean.csv').drop('Unnamed: 0',axis=1)
data = data_raw.copy(deep=True)
data_raw.info()
data_raw.describe()

## NO LONGER NEEDED
## replacing all NAN values in 'metascore' col with the existing average score
#display(data['metascore'].describe())
#display(data.info())
data['metascore'].fillna(data['metascore'].mean(), axis=0, inplace=True)
#display(data['metascore'].describe())
#display(data.info())

## Removing columns not needed for machine learning
Also removing all NaN values in the machine learning dataframe

## Remove columns not needed in machine learning
## Cols that do with year of film, name etc
data_ml = data.drop([
    'tmdb_id',
    'imdb_id',
    'film',
    'year'
], axis=1)

## remove NaN values
data_ml.dropna(axis=0, subset=['producer_0',], inplace=True)
## remove genre_id_1, producer_1 and screenplay_0, too little values, cannot be substituted for stuff like average
data_ml.drop([
    'genre_id_1',
    'producer_1',
    'screenplay_0'
], axis=1, inplace=True)
data_ml.info()
#data.head()

## Hashing objects into integers
Hashing this way allows the use of non-integer fields for machine learning

## hash all strings into ints, using sha256
strhash = lambda x: int(hashlib.sha256(x.encode('utf-8')).hexdigest(),16) % 10**8

fields = data_ml.select_dtypes('object').columns.tolist()
for field in fields:
    data_ml[field] = data_ml[field].apply(lambda x: int(hashlib.sha256(x.encode('utf-8')).hexdigest(),16) % 10**8)

data_ml.info()

## Learning using Adaptive Boosting

accuracy = {}
matrix = {}

## ML part (re-run with some headers included/ommited to see which is best)
results = data_ml['oscar_win'] ## response
movinfo = data_ml.drop(['oscar_win'#,'oscar_cat','director','producer_0','cast_3','cast_2','cast_1','cast_0'
                        ], axis=1) ## predictor, vary the columns used
display(movinfo.head(1))
## default split is 75% train, 25% test
movinfo_train, movinfo_test, results_train, results_test = train_test_split(movinfo, results) 

## Adaboost, OK
adaboost = AdaBoostClassifier(n_estimators=100, ## number of weak learners to be trained iteratively, more may be better, see results
                             learning_rate=0.2) ## something to do with how the weak learners are weighted
model = adaboost.fit(movinfo_train, results_train)

## Try out the model on test data
results_pred = model.predict(movinfo_test)
accuracy['adaboost'] = metrics.accuracy_score(results_test, results_pred)
matrix['adaboost'] = metrics.confusion_matrix(results_test, results_pred)
print("Accuracy:", accuracy['adaboost'])

## Learning using Gradient Boosting

## ML part (re-run with some headers included/ommited to see which is best) trying out gradient boosting
results = data_ml['oscar_win'] ## response
movinfo = data_ml.drop(['oscar_win'#,'oscar_cat','director','producer_0','cast_3','cast_2','cast_1','cast_0'
                        ], axis=1) ## predictor, vary the columns used
display(movinfo.head(1))
## default split is 75% train, 25% test
movinfo_train, movinfo_test, results_train, results_test = train_test_split(movinfo, results) 

## Gradboost, OK
gradbooster = GradientBoostingClassifier(n_estimators=100, ## number of weak learners to be trained iteratively, more may be better, see results
                                         learning_rate=0.2,
                                         subsample=0.6
                                        ) ## something to do with how the weak learners are weighted
model = gradbooster.fit(movinfo_train, results_train)

## Try out the model on test data
accuracy['gradboost'] = metrics.accuracy_score(results_test, results_pred)
matrix['gradboost'] = metrics.confusion_matrix(results_test, results_pred)
print("Accuracy:", accuracy['gradboost'])

## Learning using ExtRa Trees (Extremely Random Trees)

## ML part (re-run with some headers included/ommited to see which is best) trying out gradient boosting
results = data_ml['oscar_win'] ## response
movinfo = data_ml.drop(['oscar_win'#,'oscar_cat','director','producer_0','cast_3','cast_2','cast_1','cast_0'
                        ], axis=1) ## predictor, vary the columns used
display(movinfo.head(1))
## default split is 75% train, 25% test
movinfo_train, movinfo_test, results_train, results_test = train_test_split(movinfo, results) 

## ExtRaboost, OK
extratrees = ExtraTreesClassifier(n_estimators=1000, ## number of weak learners to be trained iteratively, more may be better, see results
                                  max_depth=None,
                                  min_samples_split=2,
                                  random_state=0,
                                  warm_start=True
                                ) ## something to do with how the weak learners are weighted
model = extratrees.fit(movinfo_train, results_train)

## Try out the model on test data
accuracy['extrarand'] = metrics.accuracy_score(results_test, results_pred)
matrix['extrarand'] = metrics.confusion_matrix(results_test, results_pred)
print("Accuracy:", accuracy['extrarand'])

## Learning using Random Forests

## ML part (re-run with some headers included/ommited to see which is best) trying out gradient boosting
results = data_ml['oscar_win'] ## response
movinfo = data_ml.drop(['oscar_win'#,'oscar_cat','director','producer_0','cast_3','cast_2','cast_1','cast_0'
                        ], axis=1) ## predictor, vary the columns used
display(movinfo.head(1))
## default split is 75% train, 25% test
movinfo_train, movinfo_test, results_train, results_test = train_test_split(movinfo, results) 

## RandForestboost, OK
randforest = RandomForestClassifier(n_estimators=100, ## number of weak learners to be trained iteratively, more may be better, see results
                                    warm_start=True
                                    ) ## something to do with how the weak learners are weighted
model = randforest.fit(movinfo_train, results_train)

## Try out the model on test data
accuracy['randforest'] = metrics.accuracy_score(results_test, results_pred)
matrix['randforest'] = metrics.confusion_matrix(results_test, results_pred)
print("Accuracy:", accuracy['randforest'])

## Learning using eXtreme Gradient Boosting (XGBoost)

## ML part (re-run with some headers included/ommited to see which is best) trying out gradient boosting
results = data_ml['oscar_win'] ## response
movinfo = data_ml.drop(['oscar_win'#,'oscar_cat','director','producer_0','cast_3','cast_2','cast_1','cast_0'
                        ], axis=1) ## predictor, vary the columns used
display(movinfo.head(1))
## default split is 75% train, 25% test
movinfo_train, movinfo_test, results_train, results_test = train_test_split(movinfo, results) 

num_round=100
xgb = xgboost.XGBClassifier(objective='reg:logistic',
                           max_depth = 100, 
                           alpha = 15, 
                           n_estimators = 10)
model = xgb.fit(movinfo_train, results_train)

## Try out the model on test data
accuracy['xgboost'] = metrics.accuracy_score(results_test, results_pred)
matrix['xgboost'] = metrics.confusion_matrix(results_test, results_pred)
print("Accuracy:", accuracy['xgboost'])

df_accuracy = pd.DataFrame.from_dict(accuracy, orient='index').T.reset_index(drop=True)

f = plt.figure(figsize=(12,6))
sb.barplot(data=df_accuracy)
plt.title("Comparing machine learning methods", loc='left', fontsize=15)
## comment below line out after exporting to png
#.savefig('./images/ml_acuracy_5',bbox_inches='tight',dpi=600)

f, axes = plt.subplots(1,5,figsize=(25,5))
index=0
col=0
for line in matrix:
    sb.heatmap(data=matrix[line], ax=axes[index],annot = True, fmt=".0f", annot_kws={"size": 15})
    axes[index].set_title(line, loc='left', fontsize=16)
    index+=1
plt.show()
## comment below line out after exporting to png
#f.savefig('./images/confusion_matrix_5.png',bbox_inches='tight',dpi=600)

#sb.heatmap(matrix['adaboost'],annot = True, fmt=".0f", annot_kws={"size": 15})

test_cm = metrics.confusion_matrix(results_test, results_pred)
sb.heatmap(test_cm, 
           annot = True, fmt=".0f", annot_kws={"size": 15})

## take in nominee list, parse through cleaning code
## pass thru best model(adaboost) to see the results
nom_bestactor = pd.read_csv('./data/best_actor_nominee.csv').drop('Unnamed: 0',axis=1)
nom_bestactor.info()

def makeready(df_in):
    import _functions_1115 as func
    new_headers = {
    'cast':['cast_0','cast_1','cast_2','cast_3'],
    'producers':['producer_0','producer_1'],
    'screenplay':['screenplay_0']
    }
    
    df_in.drop(['Unnamed: 0',
               'runtime',
               'nominee'], axis=1, inplace=True)
    
    df_in.rename({
    'year_film':'year',
    'oscar_category':'oscar_cat',
    'no_oscar_wins':'oscar_win_count',
    'no_oscar_nominations':'oscar_nominations',
    'staring_casts':'cast',
    'casts_popularity':'cast_popularity',
    'rotten_tomatoes_rating':'rotten_tomatoes'
    }, axis=1, inplace=True)
    
    for item in new_headers:
        df_in[item] = [func.strtolist(df_in[item][index]) for index in range(len(df_in))]
    
    for currheader in new_headers:
        ## look for target subcols, check if they exist first
        if currheader in df_in:
            ## perform addition of new cols from currheader list containing only 1 entry, NAN if list index out of range
            for subcol in range(len(new_headers[currheader])):

                tmp_list = []
                for i in range(len(df_in)):
                    ## checks if list index still in range
                    try:
                        tmp_list.append(df_in[currheader][i][subcol])
                    ## adds NAN if out of range
                    except:
                        tmp_list.append(None)

                df_in[new_headers[currheader][subcol]] = tmp_list

        df_in = df_in.convert_dtypes()
        
    df_in.drop([
    'year_ceremony','cast','producers','screenplay'
    ], axis=1, inplace=True)
    
    col_list_reorder = [
    'film',
    'budget', ## not inside?
    'revenue', ## not inside?
    'oscar_cat',
    'oscar_win',
    'oscar_nominations',
    #'oscar_win_count',
    'total_wins',
    'total_nominations',
    'cast_popularity',
    'crew_popularity',
    'tmdb_vote_average',
    'tmdb_vote_count',
    'imdb_rating',
    'imdb_votes',
    'rotten_tomatoes',
    'metascore',
    'director',
    'cast_0',
    'cast_1',
    'cast_2',
    'cast_3',
    'producer_0',
    'producer_1',
    'screenplay_0',
    'Drama',
    'History',
    'Music',
    'Action',
    'Adventure',
    'Animation',
    'Comedy',
    'Crime',
    'Family',
    'Fantasy',
    'Horror',
    'Mystery',
    'Romance',
    'Science Fiction',
    'Thriller',
    'War',
    'Western',           
    ]
    df_in = df_in[col_list_reorder]
    
    df_in.dropna(axis=0, subset=['producer_0',], inplace=True)
    ## remove genre_id_1, producer_1 and screenplay_0, too little values, cannot be substituted for stuff like average
    df_in.drop([
        'producer_1',
        'screenplay_0',
        'budget',
        'revenue'
    ], axis=1, inplace=True)
    
    return df_in

func.makeready(pd.read_csv('./data/Final_Movie_Data_Cleaned.csv'))

df = pd.read_csv('./data/best_actor_nominee.csv')
df.info()

## ML part (re-run with some headers included/ommited to see which is best)
results = data_ml['oscar_win'] ## response
movinfo = data_ml.drop(['oscar_win'#,'oscar_cat','director','producer_0','cast_3','cast_2','cast_1','cast_0'
                        ], axis=1) ## predictor, vary the columns used
display(movinfo.head(1))
## default split is 75% train, 25% test
movinfo_train, movinfo_test, results_train, results_test = train_test_split(movinfo, results) 

## Adaboost, OK
adaboost = AdaBoostClassifier(n_estimators=100, ## number of weak learners to be trained iteratively, more may be better, see results
                             learning_rate=0.2) ## something to do with how the weak learners are weighted
model = adaboost.fit(movinfo_train, results_train)

## Try out the model on test data
results_pred = model.predict(movinfo_test)
accuracy['adaboost'] = metrics.accuracy_score(results_test, results_pred)
matrix['adaboost'] = metrics.confusion_matrix(results_test, results_pred)
print("Accuracy:", accuracy['adaboost'])

---
### Test cells
---

## Try one-hot encoding of oscar award category
from sklearn.preprocessing import OneHotEncoder


## trying out encoding objs to a unique int
data_v2 = data.copy(deep=True)
data_v2_n = data_v2.drop(['producer_1','screenplay_0'], axis=1)
data_v2_n.info()

fields = [
    'oscar_cat',
    'director',
    'starring_0',
    'starring_1',
    'starring_2',
    'starring_3',
    'producer_0'
]
for field in fields:
    data_v2_n[field] = label_encoder.fit_transform(data_v2_n[field])
data_v2_n

## partitioning data into train and test
X = data_v2_n.drop(['oscar_win'], axis=1)

y = data_v2_n['oscar_win']
X_train, X_test, y_train, y_test = train_test_split(X, y)

## boost, OK
booster = AdaBoostClassifier(n_estimators=60, ## number of weak learners to be trained iteratively, more may be better, see results
                             learning_rate=0.8) ## something to do with how the weak learners are weighted

model = booster.fit(X_train, y_train)

y_pred = model.predict(X_test)
print("Accuracy:",metrics.accuracy_score(y_test, y_pred))

## Testing out hashing function to map every str to a unique int
import hashlib

def hash(sourcedf,destinationdf,*column):
    columnName = ''
    destinationdf['hash_'+columnName.join(column)] = pd.DataFrame(sourcedf[list(column)].values.sum(axis=1))[0].apply(lambda x: int(hashlib.sha512(x.encode('utf-8')).hexdigest(),16) % 10**8)

%time
## This works
strhash = lambda x: int(hashlib.sha256(x.encode('utf-8')).hexdigest(),16) % 10**8

## no longer in use
## convert all objects and strings into ints, using LabelEncoder
label_encoder = preprocessing.LabelEncoder()

fields = data_ml.select_dtypes('object').columns.tolist()
for field in fields:
    data_ml[field] = label_encoder.fit_transform(data_ml[field])
    
data_ml

z = data_ml.copy(deep=True)
fields = z.select_dtypes('object').columns.tolist()
z = z.convert_dtypes()
#z.info()
for field in fields:
    z[field] = z[field].apply(lambda x: int(hashlib.sha256(x.encode('utf-8')).hexdigest(),16) % 10**8)
#z.apply(lambda x: int(hashlib.sha256(x.encode('utf-8')).hexdigest(),16) % 10**8)
z
